services/spark_job/requirements.txt
annotated-types==0.7.0
click==8.2.1
colorama==0.4.6
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
py4j==0.10.9.9
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
services==0.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
